pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-agents-ia"
version = "1.0.0"
description = "Sistema Multi-Agentes de IA para Pesquisa e Análise usando CrewAI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Seu Nome", email = "seu.email@exemplo.com"}
]
keywords = [
    "crewai",
    "agents",
    "ai",
    "llm",
    "gemini",
    "multi-agent",
    "research",
    "automation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",
    "google-generativeai>=0.8.0",
    "litellm>=1.50.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/seu-usuario/Multi-Agents-IA"
Documentation = "https://github.com/seu-usuario/Multi-Agents-IA#readme"
Repository = "https://github.com/seu-usuario/Multi-Agents-IA.git"
Issues = "https://github.com/seu-usuario/Multi-Agents-IA/issues"

[project.scripts]
multi-agents = "main:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip = [".venv", "venv"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=. --cov-report=term-missing"
